# NodeClient/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
...
  '/api/films/public/{filmdIs}/images':
    post:
      tags:
        - films/public/{filmId}/images
      security:
        - cookieAuth: []
      summary: Associate an image to the public film
      description: >-
        The image sent in request body is associated to the public film characterized by the ID specified in the path. Only the film owner can associate an image to the film.
      operationId: addImage
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: image to assign to the film
        content:
          multipart/form-data:
            schema:
              type: object
              properties: # Request parts
                 image:
                    type: string
                    format: binary
            encoding: 
                image: 
                  contentType: image/png, image/jpg, image/gif
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '415':
          description: Unsupported Media Type
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/public/{filmId}/images/{imageId}':
    get:
      tags:
        - films/public/{filmId}/images/{imageId}
      security:
        - cookieAuth: []
      summary: Retrieve an image data structure associated to the film
      description: >-
        The image data structure to be retrieved has the ID specified in the path and it is associated to the film characterized by the ID specified in the path. Only the film owner or a film reviewer can retrieve the image data structure.
      operationId: getSingleImage
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
        - films/public/{filmId}/images/{imageId}
      security:
        - cookieAuth: []
      summary: Delete an image associated to the film
      description: >-
        The image to be deleted has the ID specified in the path and it is associated to the film characterized by the ID specified in the path. Only the film owner can delete the image.
      operationId: deleteSingleImage
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/public/{filmId}/images/{imageId}/imageFile':
    get:
      tags:
        - films/public/{filmId}/images/{imageId}/imageFile
      security:
        - cookieAuth: []
      summary: Retrieve the file of an image associated to the film
      description: >-
        The image file to be retrieved is linked to the image having the ID specified in the path and it is associated to the film characterized by the ID specified in the path. Only the film owner or a film reviewer can retrieve the image file.
      operationId: getSingleImageFile
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpg:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '414':
          description: Unsupported Media Type
          content: {}
        '500':
          description: Internal Server Error
          content: {}
components:
  schemas:
...
    Image:
      type: object
      properties:
        $schema:
          type: string
          description: schema for Image data structure
        id:
          type: integer
          description: unique identifier for the image
        name:
          type: string
          description: name without extension of the image
        fileURI:
          type: string
          format: uri-reference
          description: link to the content of the image
        self:
          type: string
          format: uri-reference
          description: self link
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid